+++ app/src/main/java/com/mongmong/namo/presentation/ui/bottom/diary/DiaryViewModel.kt
@@ -1,6 +1,5 @@
 package com.mongmong.namo.presentation.ui.bottom.diary
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -10,17 +9,11 @@ import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import androidx.paging.insertSeparators
-import com.mongmong.namo.data.local.entity.diary.Diary
 import com.mongmong.namo.data.local.entity.diary.DiarySchedule
-import com.mongmong.namo.data.local.entity.home.Schedule
-import com.mongmong.namo.domain.model.MoimDiaryResult
 import com.mongmong.namo.domain.repositories.DiaryRepository
-import com.mongmong.namo.presentation.config.RoomState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
-import java.io.File
 import javax.inject.Inject
 import org.joda.time.DateTime
 import java.text.SimpleDateFormat
@@ -29,28 +22,13 @@ import java.text.SimpleDateFormat
 class DiaryViewModel @Inject constructor(
     private val repository: DiaryRepository
 ) : ViewModel() {
-    // PersonalDetailActivity
-    private val _diary = MutableLiveData<Diary>()
-    val diary: LiveData<Diary> = _diary
-
-    private val _imgList = MutableLiveData<List<String>>(emptyList())
-    val imgList: LiveData<List<String>> = _imgList
-
-    // DiaryFragment
     private val _currentDate = MutableLiveData<String>(DateTime().toString("yyyy.MM"))
     val currentDate : LiveData<String> = _currentDate
 
     private val _isMoim = MutableLiveData<Int>(0)
     val isMoim : LiveData<Int> = _isMoim
 
-    // MoimDetailActivity
-    private val _getMoimDiaryResult = MutableLiveData<MoimDiaryResult>()
-    val getMoimDiaryResult : LiveData<MoimDiaryResult> = _getMoimDiaryResult
 
-    private val _patchDiaryResult = MutableLiveData<Boolean>()
-    val patchDiaryResult : LiveData<Boolean> = _patchDiaryResult
-
-    private val _memo = MutableLiveData<String>()
 
     /** 개인 기록 리스트 조회 **/
     fun getPersonalPaging(date: String): Flow<PagingData<DiarySchedule>> {
@@ -91,84 +69,6 @@ class DiaryViewModel @Inject constructor(
     }
     private fun Long.convertDate() : String {
         return SimpleDateFormat("yyyy.MM.dd").format(this * 1000)
-    }
-    /** 개인 기록 개별 조회 **/
-    fun getExistingPersonalDiary(diaryId: Long) {
-        viewModelScope.launch {
-            Log.d("DiaryViewModel getDiary", "$diaryId")
-            _diary.postValue(repository.getDiary(diaryId))
-        }
-    }
-    /** 개인 기록 추가시 데이터 초기화 **/
-    fun setNewPersonalDiary(schedule: Schedule, content: String) {
-        _diary.value = Diary(
-            diaryId = schedule.scheduleId,
-            scheduleServerId = schedule.serverId,
-            content = content,
-            images = _imgList.value,
-            state = RoomState.ADDED.state
-        )
-    }
-
-    /** 개인 기록 추가 **/
-    fun addPersonalDiary(images: List<File>?) {
-        viewModelScope.launch {
-            Log.d("DiaryViewModel addDiary", "$_diary")
-            _diary.value?.let {
-                repository.addDiary(
-                    diary = it,
-                    images = images
-                )
-            }
-        }
-    }
-    /** 개인 기록 수정 **/
-    fun editPersonalDiary(content: String, images: List<File>?) {
-        viewModelScope.launch {
-            _diary.value?.let {
-                it.content = content
-                it.state = RoomState.EDITED.state
-            }
-            Log.d("DiaryViewModel editDiary", "${_diary.value}")
-            _diary.value?.let {
-                repository.editDiary(
-                    diary = it,
-                    images = images
-                )
-            }
-        }
-    }
-    /** 개인 기록 삭제 **/
-    fun deletePersonalDiary(localId: Long, scheduleServerId: Long) {
-        viewModelScope.launch {
-            repository.deleteDiary(localId, scheduleServerId)
-        }
-    }
-
-    /** 모임 기록 개별 조회 **/
-    fun getMoimDiary(scheduleId: Long) {
-        viewModelScope.launch {
-            Log.d("DiaryViewModel getMoimDiary", "$scheduleId")
-            _getMoimDiaryResult.postValue(repository.getMoimDiary(scheduleId))
-        }
-    }
-
-
-    /** 모임 기록 수정 **/
-    fun patchMoimDiary(scheduleId: Long, content: String) {
-        viewModelScope.launch {
-            _patchDiaryResult.postValue(repository.patchMoimDiary(scheduleId, content))
-        }
-    }
-
-    fun setMemo(memo: String) { _memo.value = memo }
-    fun getMemo() = _memo.value
-
-
-    fun getImgList() = _imgList.value
-    fun updateImgList(newImgList: List<String>) {
-        _imgList.value = newImgList
-        _diary.value?.images = _imgList.value
     }
 
     /** 선택 날짜 **/
